pyproject.toml
[project]
name = "stateful-rag-l2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "grandalf>=0.8",
    "ipykernel>=7.2.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.1.12",
    "langchain-postgres>=0.0.17",
    "langchain-redis>=0.2.5",
    "langfuse>=4.2.0",
    "langgraph>=1.1.6",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pymupdf4llm>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
